+++ Water_Features/UI/src/mods/waterToolSections/waterToolSections.tsx
@@ -8,62 +8,76 @@ import radiusChangeSrc from "./WaterSourceChangeRadius.svg";
 import moveWaterSourceSrc from "./WaterSourceMove.svg";
 import placeWaterSourceSrc from "./WaterSourcePlacement.svg";
 import elevationChangeSrc from "./WaterSourceRaiseOrLower.svg";
+import { getModule } from "cs2/modding";
 
 // These contain the coui paths to Unified Icon Library svg assets
-export const couiStandard =                         "coui://uil/Standard/";
-export const arrowDownSrc =         couiStandard +  "ArrowDownThickStroke.svg";
-export const arrowUpSrc =           couiStandard +  "ArrowUpThickStroke.svg";
+const couiStandard =                         "coui://uil/Standard/";
+const arrowDownSrc =         couiStandard +  "ArrowDownThickStroke.svg";
+const arrowUpSrc =           couiStandard +  "ArrowUpThickStroke.svg";
 
 // These establishes the binding with C# side. Without C# side game ui will crash.
-export const AmountValue$ =        bindValue<number> (mod.id, 'AmountValue');
-export const RadiusValue$ =        bindValue<number> (mod.id, 'RadiusValue');
-export const MinDepthValue$ =      bindValue<number> (mod.id, 'MinDepthValue');
-export const AmountLocaleKey$ =    bindValue<string> (mod.id, 'AmountLocaleKey');
-export const AmountStep$ =         bindValue<number> (mod.id, 'AmountStep');
-export const RadiusStep$ =         bindValue<number> (mod.id, 'RadiusStep');
-export const MinDepthStep$ =       bindValue<number> (mod.id, 'MinDepthStep');
-export const AmountScale$ =        bindValue<number> (mod.id, 'AmountScale');
-export const RadiusScale$ =        bindValue<number> (mod.id, 'RadiusScale');
-export const MinDepthScale$ =      bindValue<number> (mod.id, 'MinDepthScale');
-export const ShowMinDepth$ =       bindValue<number> (mod.id, 'ShowMinDepth');
-export const ToolMode$ =           bindValue<number> (mod.id, "ToolMode");
+const AmountValue$ =        bindValue<number> (mod.id, 'AmountValue');
+const RadiusValue$ =        bindValue<number> (mod.id, 'RadiusValue');
+const MinDepthValue$ =      bindValue<number> (mod.id, 'MinDepthValue');
+const AmountLocaleKey$ =    bindValue<string> (mod.id, 'AmountLocaleKey');
+const AmountStep$ =         bindValue<number> (mod.id, 'AmountStep');
+const RadiusStep$ =         bindValue<number> (mod.id, 'RadiusStep');
+const MinDepthStep$ =       bindValue<number> (mod.id, 'MinDepthStep');
+const AmountScale$ =        bindValue<number> (mod.id, 'AmountScale');
+const RadiusScale$ =        bindValue<number> (mod.id, 'RadiusScale');
+const MinDepthScale$ =      bindValue<number> (mod.id, 'MinDepthScale');
+const ShowMinDepth$ =       bindValue<number> (mod.id, 'ShowMinDepth');
+const ToolMode$ =           bindValue<number> (mod.id, "ToolMode");
 
 // These are strings that will be used for translations keys and event triggers.
-export const amountDownID =             "amount-down-arrow";
-export const amountUpID =               "amount-up-arrow";
-export const radiusDownID =             "radius-down-arrow";
-export const radiusUpID =               "radius-up-arrow";
-export const minDepthDownID =           "min-depth-down-arrow";
-export const minDepthUpID =             "min-depth-up-arrow";
-export const amountStepID =             "amount-rate-of-change";
-export const radiusStepID =             "radius-rate-of-change";
-export const minDepthStepID =           "min-depth-rate-of-change";
-export const tooltipDescriptionPrefix = "YY_WATER_FEATURES_DESCRIPTION.";
-export const sectionTitlePrefix =       "YY_WATER_FEATURES.";
-export const elevationChangeID =        "ElevationChange";
-export const placeWaterSourceID =       "PlaceWaterSource";
-export const moveWaterSourceID =        "MoveWaterSource";
-export const radiusChangeID =           "RadiusChange";
+const amountDownID =             "amount-down-arrow";
+const amountUpID =               "amount-up-arrow";
+const radiusDownID =             "radius-down-arrow";
+const radiusUpID =               "radius-up-arrow";
+const minDepthDownID =           "min-depth-down-arrow";
+const minDepthUpID =             "min-depth-up-arrow";
+const amountStepID =             "amount-rate-of-change";
+const radiusStepID =             "radius-rate-of-change";
+const minDepthStepID =           "min-depth-rate-of-change";
+const tooltipDescriptionPrefix = "YY_WATER_FEATURES_DESCRIPTION.";
+const sectionTitlePrefix =       "YY_WATER_FEATURES.";
+const elevationChangeID =        "ElevationChange";
+const placeWaterSourceID =       "PlaceWaterSource";
+const moveWaterSourceID =        "MoveWaterSource";
+const radiusChangeID =           "RadiusChange";
 
 // Stores the default values for the step arrays. Must be descending order.
-export const defaultValues : number[] =[1.0, 0.5, 0.25, 0.125];
+const defaultValues : number[] =[1.0, 0.5, 0.25, 0.125];
 
 // This functions trigger an event on C# side and C# designates the method to implement.
-export function handleClick(eventName: string) {
+function handleClick(eventName: string) {
     trigger(mod.id, eventName);
 }
 
 // This function triggers an event to change the water tool mode to specified tool mode.
-export function changeToolMode(toolMode: WaterToolModes) {
+function changeToolMode(toolMode: WaterToolModes) {
     trigger(mod.id, "ChangeToolMode", toolMode);
 }
 
-export enum WaterToolModes 
+enum WaterToolModes 
 {
     PlaceWaterSource,
     ElevationChange,
     MoveWaterSource,
     RadiusChange,
+}
+
+const descriptionToolTipStyle = getModule("game-ui/common/tooltip/description-tooltip/description-tooltip.module.scss", "classes");
+    
+
+// This is working, but it's possible a better solution is possible.
+export function descriptionTooltip(tooltipTitle: string | null, tooltipDescription: string | null) : JSX.Element {
+    return (
+        <>
+            <div className={descriptionToolTipStyle.title}>{tooltipTitle}</div>
+            <div className={descriptionToolTipStyle.content}>{tooltipDescription}</div>
+        </>
+    );
 }
 
 export const WaterToolComponent: ModuleRegistryExtend = (Component : any) => {
@@ -108,6 +122,11 @@ export const WaterToolComponent: ModuleRegistryExtend = (Component : any) => {
         const moveWaterSourceTooltip =  translate(tooltipDescriptionPrefix + moveWaterSourceID, "Water Tool will move existing water sources. Target elevations of existing water sources will not change. Right click to cancel.");
         const radiusChangeTooltip =     translate(tooltipDescriptionPrefix + radiusChangeID,    "Water Tool will change radius of water sources. Right click to cancel.");
         const toolModeTitle =           translate("Toolbar.TOOL_MODE_TITLE", "Tool Mode");
+        const placeWaterSourceTitle =   translate(sectionTitlePrefix + placeWaterSourceID,      "Place Water Source");
+        const moveWaterSourceTitle =    translate(sectionTitlePrefix + moveWaterSourceID,       "Move Water Source");
+        const elevationChangeTitle =    translate(sectionTitlePrefix + elevationChangeID,       "Elevation Change");
+        const radiusChangeTitle =       translate(sectionTitlePrefix + radiusChangeID,          "Change Radius");
+
 
         var result = Component();
         if (toolActive) 
@@ -182,10 +201,10 @@ export const WaterToolComponent: ModuleRegistryExtend = (Component : any) => {
                         </VanillaComponentResolver.instance.Section>
                     )}
                     <VanillaComponentResolver.instance.Section title={toolModeTitle}>
-                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.PlaceWaterSource}    tooltip={placeWaterSourceTooltip}  onSelect={() => changeToolMode(WaterToolModes.PlaceWaterSource)}   src={placeWaterSourceSrc} focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.ElevationChange}     tooltip={elevationChangeTooltip}   onSelect={() => changeToolMode(WaterToolModes.ElevationChange)}    src={elevationChangeSrc}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.MoveWaterSource}     tooltip={moveWaterSourceTooltip}   onSelect={() => changeToolMode(WaterToolModes.MoveWaterSource)}    src={moveWaterSourceSrc}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.RadiusChange}        tooltip={radiusChangeTooltip}      onSelect={() => changeToolMode(WaterToolModes.RadiusChange)}       src={radiusChangeSrc}     focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.PlaceWaterSource}    tooltip={descriptionTooltip(placeWaterSourceTitle, placeWaterSourceTooltip)}  onSelect={() => changeToolMode(WaterToolModes.PlaceWaterSource)}   src={placeWaterSourceSrc} focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.ElevationChange}     tooltip={descriptionTooltip(elevationChangeTitle, elevationChangeTooltip)}     onSelect={() => changeToolMode(WaterToolModes.ElevationChange)}    src={elevationChangeSrc}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.MoveWaterSource}     tooltip={descriptionTooltip(moveWaterSourceTitle, moveWaterSourceTooltip)}     onSelect={() => changeToolMode(WaterToolModes.MoveWaterSource)}    src={moveWaterSourceSrc}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={ToolMode == WaterToolModes.RadiusChange}        tooltip={descriptionTooltip(radiusChangeTitle, radiusChangeTooltip)}           onSelect={() => changeToolMode(WaterToolModes.RadiusChange)}       src={radiusChangeSrc}     focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
                     </VanillaComponentResolver.instance.Section>                    
                 </>
             );