+++ Water_Features/ChangeLog.md
@@ -1,3 +1,10 @@
-﻿## Patch V1.1.1
-* New icons by Chameleon TBN
-* Cleanup some React UI code.
+﻿## Patch V1.1.2
+* Fixed Editor water panel removes all water sources placed with Yenyang's custom water tool.
+* Fixed water sources placed in editor with Yenyang's custom water tool do not appear in Editor water panel.
+* Changed lakes placed in editor will behave per Constant Level Water Sources in editor.
+* Updated localization.
+* Added tooltip titles for tool modes.
+* Added basic tooltips with titles to water source prefabs in editor.
+* Potentially fixed settings bug that preventing settings from saving occasionally.
+* Reduced step for Seasonal Streams -> Constant Flow Rate from 5 to 1.
+* Fixed issue with experimental setting for Waves and Tides in Editor was not handled correctly.